fix: make irpc-iroh compile in wasm, and check in CI (#66)
`irpc-iroh` does not compile to wasm atm because we imported the wrong
BoxFuture alias

(in n0_future there are two: one that has a Send bound always, and one
that has one on non-wasm targets and does not have one on wasm targets.
We might want to revisit this).

Also we only checked irpc but not irpc-iroh in CI for wasm compilation.
Fixed here too.