refactor: Modify error types (#69)
mpsc::Receiver::recv returns a Result<Option<T>, RecvError>. RecvError
has a case `SenderClosed`. So we have 2 different ways to report that
the remote has closed the sender normally. Ok(None) and
Err(RecvError::SenderClosed). This is confusing. SenderClosed isn't
really an error, it is just what you get if the other side closes the
stream.

So this PR creates separate RecvError for oneshot and mpsc.

In oneshot we still need the SenderClosed case since we expect exactly 1
item, and therefore the receiver being closed before that item is an
error.

In mpsc we remove the SenderClosed error and just report Ok(None) in
that case.