feat: varint feature flag (#54)
* iroh-blobs uses the varint utils. Currently they are behind the `rpc`
feature. However iroh-blobs would use those even when the quinn rpc is
disabled. So I added a feature flag for this. Maybe the better thing
would be to move those out into a varint crate? Dunno. But this is fine
too.
* Added a `RequestError::Unreachable` variant if `rpc` feature is
disabled. `Snafu` chokes on empty enums somehow. This was the quick fix
to have `iroh_blobs` build without irpc's `rpc` feature. Not sure if
there is a better way, would need to investigate what exactly snafu
does.